+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Unreleased
 
+- Handle array for assign responseBodyProp 
 - Documented  "operationPreRequestScripts" properties in the readme
 
 ## v1.8.1 - (2021-09-02)